+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.2] - 2026-07-12
+
+### Fixed
+
+- Copa Classic races now use modern dates instead of the historical model year
+  of one of the cars.
+- Copa Uno races now use dates from the 1990s.
+
 ## [1.8.1] - 2026-07-12
 
 ### Fixed